feat(version): Pin to specific versions
I'm trying to figure out the best way to have this container pinned to a
specific version of Atuin so the build are reproducible. Otherwise it
would not check out the `cached` stuff from GitHub and you'd end up
building wihtout cache. Annoying.

Eventually this should be made so it just builds and pushes tagged
versions of the container, but baby steps along the way.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>